Change to Transcript Result Label
US3-5

Before, when rerecording, the result of the 2nd attempt wouldn't show up. Basically fixed the Label l, where before we created a new label each time we pressed stop record. This is bad, as it would not only keep the old label l and display the old result, but also because we create 2 instances of the same name, causes errors. Thus, made it so we construct l and set it up in the constructor, but change the text each time stopRecord called.