CreateHandler Made
US3-8
US3-Story Test
US4-6
US4-Story Test

Before, we had a Recipe that we would pass along the various UI pages to keep track of info, but that violated the 3 Tier architecture. Thus, we made CreateHandler, which itself contains a Recipe that we can edit along, thus having the same functionality while also ensuring that 3 tier architecture is preserved.