Recording Uploaded to Server
Changed the project code so now our recorder will first record a recording.wav in the client/View (local), then will call a performAudioRequest that does a PUT command to put the recording.wav in the Server/Model (server). Within that, it will also call the Whisper.transcribe, which itself does a POST request to WhisperAPI and returns the transcript result from performAudioRequest.