Generate Info Uploaded to Server
Changed the project code so now our recorder will call the performGenerateRequest that does a PUT command to put the mealtype and ingredients in the Server/Model (server). Within that, it will also call the generate, which itself does a POST request to ChatGPT and returns the recipe (including instructions and a title) result to performGenerateRequest.